backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
twilio